netlink: Fix dump skb leak/double free
When we free cb->skb after a dump, we do it after releasing the
lock.  This means that a new dump could have started in the time
being and we'll end up freeing their skb instead of ours.

This patch saves the skb and module before we unlock so we free
the right memory.

Fixes: 16b304f3404f ("netlink: Eliminate kmalloc in netlink dump operation.")
Reported-by: Baozeng Ding <sploving1@gmail.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Acked-by: Cong Wang <xiyou.wangcong@gmail.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Git-repo:http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git
Git-commit: 92964c79b357efd980812c4de5c1fd2ec8bb5520

Change-Id: If2ea32cfd98906c8bce71dee4c086a24a58a274f
Signed-off-by: Srinivasa Rao Kuppala <srkupp@codeaurora.org>